Move import_single_block to root
