+++ core/consensus/common/src/import_queue.rs
@@ -25,11 +25,11 @@
 //! instantiated. The `BasicQueue` and `BasicVerifier` traits allow serial
 //! queues to be instantiated simply.
 
-use std::sync::Arc;
-use runtime_primitives::{Justification, traits::{Block as BlockT, NumberFor}};
+use std::{sync::Arc, collections::HashMap};
+use runtime_primitives::{Justification, traits::{Block as BlockT, Header as _, NumberFor}};
 use crate::{error::Error as ConsensusError, well_known_cache_keys::Id as CacheKeyId};
 use crate::block_import::{
-	BlockImport, BlockOrigin, ImportBlock, ImportedAux, JustificationImport,
+	BlockImport, BlockOrigin, ImportBlock, ImportedAux, JustificationImport, ImportResult,
 	FinalityProofImport, FinalityProofRequestBuilder,
 };
 
@@ -172,4 +172,76 @@ pub enum BlockImportError {
 	UnknownParent,
 	/// Other Error.
 	Error,
+}
+
+/// Single block import function.
+pub fn import_single_block<B: BlockT, V: Verifier<B>>(
+	import_handle: &dyn BlockImport<B, Error = ConsensusError>,
+	block_origin: BlockOrigin,
+	block: IncomingBlock<B>,
+	verifier: Arc<V>,
+) -> Result<BlockImportResult<NumberFor<B>>, BlockImportError> {
+	let peer = block.origin;
+
+	let (header, justification) = match (block.header, block.justification) {
+		(Some(header), justification) => (header, justification),
+		(None, _) => {
+			if let Some(ref peer) = peer {
+				debug!(target: "sync", "Header {} was not provided by {} ", block.hash, peer);
+			} else {
+				debug!(target: "sync", "Header {} was not provided ", block.hash);
+			}
+			return Err(BlockImportError::IncompleteHeader(peer))
+		},
+	};
+
+	trace!(target: "sync", "Header {} has {:?} logs", block.hash, header.digest().logs().len());
+
+	let number = header.number().clone();
+	let hash = header.hash();
+	let parent = header.parent_hash().clone();
+
+	let import_error = |e| {
+		match e {
+			Ok(ImportResult::AlreadyInChain) => {
+				trace!(target: "sync", "Block already in chain {}: {:?}", number, hash);
+				Ok(BlockImportResult::ImportedKnown(number))
+			},
+			Ok(ImportResult::Imported(aux)) => Ok(BlockImportResult::ImportedUnknown(number, aux, peer.clone())),
+			Ok(ImportResult::UnknownParent) => {
+				debug!(target: "sync", "Block with unknown parent {}: {:?}, parent: {:?}", number, hash, parent);
+				Err(BlockImportError::UnknownParent)
+			},
+			Ok(ImportResult::KnownBad) => {
+				debug!(target: "sync", "Peer gave us a bad block {}: {:?}", number, hash);
+				Err(BlockImportError::BadBlock(peer.clone()))
+			},
+			Err(e) => {
+				debug!(target: "sync", "Error importing block {}: {:?}: {:?}", number, hash, e);
+				Err(BlockImportError::Error)
+			}
+		}
+	};
+
+	match import_error(import_handle.check_block(hash, parent))? {
+		BlockImportResult::ImportedUnknown { .. } => (),
+		r => return Ok(r), // Any other successful result means that the block is already imported.
+	}
+
+	let (import_block, maybe_keys) = verifier.verify(block_origin, header, justification, block.body)
+		.map_err(|msg| {
+			if let Some(ref peer) = peer {
+				trace!(target: "sync", "Verifying {}({}) from {} failed: {}", number, hash, peer, msg);
+			} else {
+				trace!(target: "sync", "Verifying {}({}) failed: {}", number, hash, msg);
+			}
+			BlockImportError::VerificationFailed(peer.clone(), msg)
+		})?;
+
+	let mut cache = HashMap::new();
+	if let Some(keys) = maybe_keys {
+		cache.extend(keys.into_iter());
+	}
+
+	import_error(import_handle.import_block(import_block, cache))
 }
